+++ assets/utils/offering.ts
@@ -1,0 +1,22 @@
+import { OfferingTypeEnum } from "@/@types/app";
+
+export const offeringTypeDisplayName = (type: OfferingTypeEnum): string => {
+    switch (type) {
+        case OfferingTypeEnum.WMSVECTOR:
+            return "Web Map Service Vecteur";
+        case OfferingTypeEnum.WFS:
+            return "Web Feature Service";
+        case OfferingTypeEnum.WMTSTMS:
+            return "Web Map Tile Service - Tile Map Service";
+        case OfferingTypeEnum.WMSRASTER:
+            return "Web Map Service Raster";
+        case OfferingTypeEnum.DOWNLOAD:
+            return "Service de Téléchargement";
+        case OfferingTypeEnum.ITINERARYISOCURVE:
+            return "Service de calcul d’itinéraire / isochrone";
+        case OfferingTypeEnum.ALTIMETRY:
+            return "Service d’altimétrie";
+        default:
+            return type;
+    }
+};
